backend/image-upload/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
billiard==4.2.3
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.8.0
dotenv==0.9.9
fastapi==0.121.3
h11==0.16.0
idna==3.11
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
redis==5.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
